Change "Drake's Request" to "Drake's Command" on party sheet
[R1 Errata] The second item in the Special Conditions for Opening Envelopes in the rule book talks about Drake's Request but that isn't referenced in the Scenario Book anywhere. Is this an error?.
Yes. The second item should read "The Drake's Command" instead of "The Drake's Request." When you achieve this goal, you earn "The Drake Aided" global achievement. Also, it is optional to add the cards to the event decks and earn the achievement, but once it is done, it is permanent.